YouTube: Unable to move the volume slider by dragging after disabling Video Viewer https://bugs.webkit.org/show_bug.cgi?id=276496 rdar://128587639
Reviewed by Eric Carlson.

When entering fullscreen, we add a dragstart event listener to the window that we never remove.
This makes it so that after we have exited fullscreen, the dragstart event listener is still
active, and it has a side effect of making the elements on the page to be draggable. This is
noticeable when the user is moving youtube's volume slider. Instead of this moving the knob on
the slider, this drags an image of the slider.

To fix this, we should remove the dragstart event listener on the window when we exit fullscreen.
This will remove the event listener's side effect that is causing this bug.

* Source/WebCore/Modules/modern-media-controls/controls/media-controls.js:
(MediaControls.prototype.disable):
(MediaControls):
* Source/WebCore/Modules/modern-media-controls/media/media-controller.js:
(MediaController.prototype._updateControlsIfNeeded):

Canonical link: https://commits.webkit.org/280873@main